+++ quick-start.sh
@@ -3,8 +3,8 @@ set -o errexit
 
 pip install -r requirements.txt > /dev/null
 
-mkdir instance
-echo 'You will need to update app.conf in order to login with Globus.'
-cp example-app.conf instance/app.conf
+[ -d instance ] || \
+  mkdir instance; cp example-app.conf instance/app.conf; \
+  echo 'You will need to update app.conf in order to login with Globus.'
 
 FLASK_ENV=development ./app.py