+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## in progress
+### Added
+- Globus login/logout
+
 ## [0.0.2](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.2) - 2019-10-30
 ### Added
 - Start applying Bootstrap styles.